DC: improved calculation;

+++ internal/model/calculation.go
@@ -1,7 +1,7 @@
 package model
 
 type CalcState struct {
-	PreviousNum float32 `json:"previousNum"`
-	CurrentNum  float32 `json:"currentNum"`
+	PreviousNum float64 `json:"previousNum"`
+	CurrentNum  float64 `json:"currentNum"`
 	Operation   string  `json:"operation"`
 }
